+++ src/YesSql.Core/Indexes/IdentityMap.cs
@@ -4,16 +4,16 @@ namespace YesSql.Indexes
 {
     public class IdentityMap
     {
-        private readonly Dictionary<int, object> _documentIds = new Dictionary<int, object>();
-        private readonly Dictionary<object, int> _entities = new Dictionary<object, int>();
-        private readonly Dictionary<int, Document> _documents = new Dictionary<int, Document>();
+        private readonly Dictionary<long, object> _documentIds = new Dictionary<long, object>();
+        private readonly Dictionary<object, long> _entities = new Dictionary<object, long>();
+        private readonly Dictionary<long, Document> _documents = new Dictionary<long, Document>();
 
-        public bool TryGetDocumentId(object item, out int id)
+        public bool TryGetDocumentId(object item, out long id)
         {
             return _entities.TryGetValue(item, out id);
         }
 
-        public bool TryGetEntityById(int id, out object document)
+        public bool TryGetEntityById(long id, out object document)
         {
             return _documentIds.TryGetValue(id, out document);
         }
@@ -23,7 +23,7 @@ namespace YesSql.Indexes
             return _entities.ContainsKey(entity);
         }
 
-        public void AddEntity(int id, object entity)
+        public void AddEntity(long id, object entity)
         {
             _entities.Add(entity, id);
             _documentIds.Add(id, entity);
@@ -34,12 +34,12 @@ namespace YesSql.Indexes
             _documents[doc.Id] = doc;
         }
 
-        public bool TryGetDocument(int id, out Document doc)
+        public bool TryGetDocument(long id, out Document doc)
         {
             return _documents.TryGetValue(id, out doc);
         }
 
-        public void Remove(int id, object entity)
+        public void Remove(long id, object entity)
         {
             _entities.Remove(entity);
             _documentIds.Remove(id);
